+++ back-end/src/main/java/com/obligatorio/obligatorio2dda/service/PlanesViajeService.java
@@ -11,8 +11,8 @@ import com.obligatorio.obligatorio2dda.entity.PlanesViaje;
 @Service
 public interface PlanesViajeService {
     public Iterable<PlanesViaje> findAll();
-    public Optional<PlanesViaje> findById(Long id);
+    public Optional<PlanesViaje> findById(int id);
     public PlanesViaje save(PlanesViaje save);
-    public void deleteById(Long Id);
+    public void deleteById(int Id);
     public Iterable<PlanesViaje> getPlansListing(ArrayList<Integer> ids, String fecha);
 }
